Combine js file output into one file, and etc
-Picker undefine error fixed
-CreateBoundingBox() making BoxHelper
-Scenes member function type and constructor contents changed
-BuildScene code Move to LoadModelManager

and Optimizing the framework

+++ JWFramework/Picker/Picker.ts
@@ -1,6 +1,9 @@
-﻿namespace JWFramework {
-    export class Picker {
-        public constructor() {
+﻿namespace JWFramework
+{
+    export class Picker
+    {
+        public constructor()
+        {
             this.raycaster = new THREE.Raycaster();
             this.pickedObject = null;
 
@@ -8,22 +11,27 @@
 
             this.CreateOrtbitControl();
 
-            window.addEventListener('mousemove', function (e) {
+            window.addEventListener('mousemove', function (e)
+            {
                 SceneManager.getInstance().CurrentScene.Picker.mouseEvent = e;
             });
 
-            window.addEventListener('click', function (e) {
+            window.addEventListener('click', function (e)
+            {
                 SceneManager.getInstance().CurrentScene.Picker.SetPickPosition(e);
             });
-            window.addEventListener('mouseout', function (e) {
+            window.addEventListener('mouseout', function (e)
+            {
                 SceneManager.getInstance().CurrentScene.Picker.ClearPickPosition();
             });
-            window.addEventListener('mouseleave', function (e) {
+            window.addEventListener('mouseleave', function (e)
+            {
                 SceneManager.getInstance().CurrentScene.Picker.ClearPickPosition();
             });
         }
 
-        private CreateOrtbitControl() {
+        private CreateOrtbitControl()
+        {
             this.orbitControl = new THREE.OrbitControls(WorldManager.getInstance().MainCamera.CameraInstance, WorldManager.getInstance().Canvas);
             this.orbitControl.maxDistance = 2000;
             this.orbitControl.minDistance = -2000;
@@ -32,17 +40,24 @@
             this.orbitControl.panSpeed = 3;
         }
 
-        private GetParentName(intersectedObjects: THREE.Object3D) {
-            if (intersectedObjects.type != "Group") {
-                this.GetParentName(intersectedObjects.parent);
-            }
-            else {
-                this.pickedParentName = intersectedObjects.name;
-                this.pickedObject = intersectedObjects;
+        private GetParentName(intersectedObjects: THREE.Object3D)
+        {
+            if (intersectedObjects != null) {
+                if (intersectedObjects.type == "Mesh") {
+                    this.pickedParentName = undefined;
+                }
+                if (intersectedObjects.type != "Group") {
+                    this.GetParentName(intersectedObjects.parent);
+                }
+                else {
+                    this.pickedParentName = intersectedObjects.name;
+                    this.pickedObject = intersectedObjects;
+                }
             }
         }
 
-        private PickOffObject() {
+        private PickOffObject()
+        {
             let objectList = ObjectManager.getInstance().GetObjectList;
             for (let TYPE = ObjectType.OBJ_OBJECT3D; TYPE < ObjectType.OBJ_OBJECT2D; ++TYPE) {
                 for (let OBJ = 0; OBJ < objectList[TYPE].length; ++OBJ) {
@@ -51,8 +66,8 @@
             }
         }
 
-        public Pick() {
-
+        public Pick()
+        {
             let terrain;
 
             if (this.pickPositionX > 0.75 || this.pickPositionX == -1) {
@@ -82,15 +97,17 @@
             }
             //터레인은 키보드 입력으로 높낮이 조절 가능하게 할것
             else if (this.pickMode == PickMode.PICK_TERRAIN) {
+                GUIManager.getInstance().GUI_Terrain.SetTerrainOptionList();
+                let heightOffset = GUIManager.getInstance().GUI_Terrain.GetHeightOffset();
                 let objectManager = ObjectManager.getInstance();
                 let intersectedObject = this.raycaster.intersectObject(SceneManager.getInstance().SceneInstance, true);
                 if (intersectedObject[0] != undefined) {
                     terrain = objectManager.GetObjectFromName(intersectedObject[0].object.name);
                     if (terrain != null && terrain.Type == ObjectType.OBJ_TERRAIN) {
                         intersectedObject[0].face.normal;
-                        (terrain as unknown as HeightmapTerrain).SetHeight(intersectedObject[0].face.a, 1, this.terrainOption);
-                        (terrain as unknown as HeightmapTerrain).SetHeight(intersectedObject[0].face.b, 1, this.terrainOption);
-                        (terrain as unknown as HeightmapTerrain).SetHeight(intersectedObject[0].face.c, 1, this.terrainOption);
+                        (terrain as unknown as HeightmapTerrain).SetHeight(intersectedObject[0].face.a, heightOffset, GUIManager.getInstance().GUI_Terrain.GetTerrainOption());
+                        (terrain as unknown as HeightmapTerrain).SetHeight(intersectedObject[0].face.b, heightOffset, GUIManager.getInstance().GUI_Terrain.GetTerrainOption());
+                        (terrain as unknown as HeightmapTerrain).SetHeight(intersectedObject[0].face.c, heightOffset, GUIManager.getInstance().GUI_Terrain.GetTerrainOption());
                     }
                 }
                 //SceneManager.getInstance().SceneInstance.add(objectManager.GetInSectorTerrain());
@@ -101,7 +118,8 @@
                     this.GetParentName(intersectedObjects[0].object);
                     this.pickedParent = ObjectManager.getInstance().GetObjectFromName(this.pickedParentName);
                     console.log(this.pickedParentName);
-                    this.pickedParent.DeleteObject();
+                    if (this.pickedParentName != undefined)
+                        this.pickedParent.DeleteObject();
                 }
             }
             else {
@@ -110,69 +128,73 @@
                     this.GetParentName(intersectedObjects[0].object);
                     this.pickedParent = ObjectManager.getInstance().GetObjectFromName(this.pickedParentName);
                     console.log(this.pickedParentName);
-                    this.pickedParent.Picked = true;
-                    GUIManager.getInstance().GUI_SRT.SetGameObject(this.pickedParent);
+                    if (this.pickedParentName != undefined) {
+                        this.pickedParent.Picked = true;
+                        GUIManager.getInstance().GUI_SRT.SetGameObject(this.pickedParent);
+                    }
                 }
             }
         }
 
-        private GetCanvasReleativePosition(event) {
+        private GetCanvasReleativePosition(event)
+        {
             let rect = WorldManager.getInstance().Canvas.getBoundingClientRect();
             return {
                 x: (event.clientX - rect.left) * WorldManager.getInstance().Canvas.width / rect.width,
-                y: (event.clientY - rect.top) * WorldManager.getInstance().Canvas.height / rect.height, 
+                y: (event.clientY - rect.top) * WorldManager.getInstance().Canvas.height / rect.height,
             };
         }
 
-        public get MouseEvent(): MouseEvent {
+        public get MouseEvent(): MouseEvent
+        {
             return this.mouseEvent;
         }
 
-        public SetPickPosition(event) {
+        public SetPickPosition(event)
+        {
             this.pickPositionX = (event.clientX / window.innerWidth) * 2 - 1;
             this.pickPositionY = - (event.clientY / window.innerHeight) * 2 + 1;
             this.Pick();
         }
 
-        public ClearPickPosition() {
+        public ClearPickPosition()
+        {
             this.pickPositionX = -10000;
             this.pickPositionY = -10000;
         }
 
-        public ChangePickModeModify() {
+        public ChangePickModeModify()
+        {
             this.pickMode = PickMode.PICK_MODIFY;
         }
 
-        public ChangePickModeClone() {
+        public ChangePickModeClone()
+        {
             this.pickMode = PickMode.PICK_CLONE;
         }
 
-        public ChangePickModeTerrain() {
+        public ChangePickModeTerrain()
+        {
             this.pickMode = PickMode.PICK_TERRAIN;
         }
 
-        public ChangePickModeRemove() {
+        public ChangePickModeRemove()
+        {
             this.pickMode = PickMode.PICK_REMOVE;
         }
 
-        public ChangeTerrainOption() {
-            if (this.pickMode == PickMode.PICK_TERRAIN) {
-                if (this.terrainOption == TerrainOption.TERRAIN_BALANCE)
-                    this.terrainOption = TerrainOption.TERRAIN_UP;
-                else
-                    this.terrainOption++;
-            }
-        }
-
-        public get PickMode(): PickMode {
+        public get PickMode(): PickMode
+        {
             return this.pickMode;
         }
 
-        public get OrbitControl(): THREE.OrbitControls {
+        public get OrbitControl(): THREE.OrbitControls
+        {
             return this.orbitControl;
         }
 
-        public GetPickParents(): GameObject {
+        public GetPickParents(): GameObject
+        {
             return this.pickedParent;
         }
 
@@ -185,9 +207,6 @@
         private pickedParent: GameObject;
         private pickPositionX = 0;
         private pickPositionY = 0;
-
-        private terrainOption: TerrainOption = TerrainOption.TERRAIN_UP;
-        private terrainOffset: number = 0;
 
         private orbitControl: THREE.OrbitControls;
